add documentation, rename functions, move util func, clean up issues
