+++ app/src/Modals/KYC/steps/Nationality.js
@@ -6,7 +6,7 @@ import style from "../kyc.scss";
 import Select from "components/Form/Select";
 import Spinner from "components/Spinner";
 
-import { getResidenceCountries } from "api/whitelist";
+import { getResidenceCountries } from "api/onboarding";
 
 import { STEP_REJECTED, STEP_PERSONAL } from "../";
 import { isRequired, validator } from "utils/validations";
