+++ app/src/Modals/KYC/steps/Pending.js
@@ -7,7 +7,7 @@ import useGlobalState from "hooks/useGlobalState";
 
 import useInterval from "@use-it/interval";
 
-import { getWhitelistSddIsProcessing } from "api/whitelist";
+import { isTieredWhitelistProcessing } from "api/onboarding";
 import { STEP_REJECTED, STEP_APPROVED } from "..";
 
 const cx = classnames.bind(style);
@@ -25,7 +25,7 @@ const Pending = ({ closeModal, handleAdvanceStep }) => {
     }
 
     return (async () => {
-      const status = await getWhitelistSddIsProcessing(account);
+      const status = await isTieredWhitelistProcessing(account);
 
       if (
         status["sanctionStatus"] === "ERROR" ||
