+++ app/src/api/util.js
@@ -1,0 +1,9 @@
+export const createQueryString = params => {
+  return Object.keys(params)
+    .map(key => {
+      if (params[key]) {
+        return key + "=" + params[key];
+      }
+    })
+    .join("&");
+};
